Begin Functional Programming
Begin new section, Functional Programming, in which we learn how to build functions which are independent, not affecting global state, without side effects on data objects. We will use callbacks, which are functions provided to other functions to decide an outcome. All functions are first order functions. Higher order functions accept functions as arguments, and may return functions; these argument and return functions are known as lambas. Or is it lambdae?