Solve the remove Window tabs problem using splice
Solve the remove Windows tabs problem using splice. The original solution did not appear to be using splice correctly. Splice returns the removed items. The first parameter is the index to begin removing. The second parameter is the quantity of items to remove. In this case, we only want to remove one item, at the index, which is provided as an argument