Learn how to use map array method
Learn how to use the array method, map, which can be used to replace for loops. The map method receives up to three arguments, the first being the current element of the array on which it was enacted. The map method uses a callback function which acts on each element of the array. It returns the array after it has had the callback enacted on it.