Add more information about the map method
Describe how the map method works, in terms of what it does with the provided array. It does not mutate the global variable represented by the array parameter; it returns a same-length array which has been worked on upon by the callback function. Do you like my long commit messages, or am I just sending this information into a void?