Add notes about sort
Add notes about how sort works. It expects a callback function to compare the two parameters, a and b, which represents two values in the array. If the comparison of a and b returns a number which is greater than 0, then b goes before a. Otherwise, the array order is not affected. This appears to be a mutating method.