Solve sum-array-nums
Declare a reducer function. Map the input array, then sort it. Remove the first and last number, using them in a for loop as the first index and the ending condition, respectively. Add each 1 increment to the local array. Next, use the reducer on the local array, providing an initial value of 0 to the current number. For each value in the local array, the reduce method will add it to 0, returning one single value, which is the sum of the range.