Solve sum of nums using recursion
Use destructuring on the input array to get the first and second values, except make sure to declare a default array just
in case someone passes undefined into your function. Get the increment value by checking whether the difference between the first and second numbers is less than 0, setting the increment to 1 or -1 accordingly. Finally, checking whether the first number is not equal to the last number; when false, meaning equality, then pass the first value.