Solve wherefore-art-thou clumsily
Use for loop on collection, Object.keys on source, then loop through each key in the source to check whether the collection has that property AND that the value at that property is the same as in source. If anything fails, push on an arbitrary value to a mismtaches array. Before going on to the next iteration of the for loop, check whether the length of mismatches is 0; if yes, then push the current collection object onto the return array. The mismatches array will reinitialize each time that the for loop runs. Same with keysToCheck, which is unnecessary, so it should be moved outside the for loop.