Simplfy whatIsInAName to use filter.
Creating an arbitrary array, then pushing on an arbitrary value, just to sort another array, was inefficient. Instead, create a variable to store the Array result of Object.keys(source). Next, return the array result from filtering the collection. Perform a loop on the sourceKeys array, preventing any Objects in the collection being returned if it doesn't have the same property or value as the source.